Move string parsing stuff to `SecureExtensions`
