+++ UnityProject/Assets/Scripts/US13/Core/SecureStuff/AllowedReflection.cs
@@ -4,7 +4,9 @@ using System.Globalization;
 using System.Linq;
 using System.Reflection;
 using Logs;
+using Mirror.BouncyCastle.Asn1.X509.Qualified;
 using NUnit.Framework;
+using SecureStuff.Util;
 using UnityEngine;
 using UnityEngine.Events;
 
@@ -450,32 +452,48 @@ namespace SecureStuff
 		}
 
 		/// <summary>
-		/// Sets the value of a property or field on the specified object using reflection.
-		/// The target member can be either a writable property or a field. The provided value
-		/// is converted to the target member's type before assignment.
+		/// Sets the value of a property or field on the specified object if the provided value
+		/// is compatible with the member's type.
 		/// </summary>
-		/// <param name="targetVariable">The name of the property or field to update on the target object.</param>
-		/// <param name="valueToChange">The value to assign to the specified property or field. The value will be converted</param>
-		/// <param name="targetObject">The object instance containing the property or field to modify.</param>
-		/// <returns><c>true</c> if the property or field was found and successfully updated;
-		/// otherwise, <c>false</c>.</returns>
+		/// <param name="targetVariable">
+		/// The name of the property or field to update.
+		/// </param>
+		/// <param name="valueToChange">
+		/// The value to assign. The value must already be of a compatible type.
+		/// </param>
+		/// <param name="targetObject">
+		/// The object containing the property or field to modify.
+		/// </param>
+		/// <returns>
+		/// <c>true</c> if the member was found and updated; otherwise, <c>false</c>.
+		/// </returns>
 		public static bool SetVariable(string targetVariable, object valueToChange, object targetObject)
 		{
+			if (targetObject == null) return false;
+			if (valueToChange == null) return false;
+
 			var type = targetObject.GetType();
+
 			var property = type.GetProperty(targetVariable);
 			if (property != null && property.CanWrite)
 			{
-				property.SetValue(targetObject, Convert.ChangeType(valueToChange, property.PropertyType));
+				if (property.PropertyType.IsInstanceOfType(valueToChange) == false) return false;
+				property.SetValue(targetObject, valueToChange);
 				return true;
+
 			}
 
 			var field = type.GetField(targetVariable);
-			if (field != null)
+			if (field == null) return false;
+			if (field.FieldType.IsInstanceOfType(valueToChange) == false)
 			{
-				field.SetValue(targetObject, Convert.ChangeType(valueToChange, field.FieldType));
-				return true;
+				if (valueToChange.GetType() != typeof(string)) return false;
+				string s = (string)valueToChange;
+				Tuple<Type, object> parsedValue = s.ParseString();
+				if (parsedValue.Item2 == null || parsedValue.Item1 != field.FieldType) return false;
+				field.SetValue(targetObject, parsedValue.Item2);
 			}
-			return false;
+			return true;
 		}
 	}
 }
